requirements.txt
Django==1.7
Pillow==2.5.3
boto==2.34.0
dj-database-url==0.3.0
dj-static==0.0.6
django-bootstrap3==4.11.0
django-boto==0.3.8
django-storages==1.1.8
django-toolbelt==0.0.1
gunicorn==19.1.1
psycopg2==2.5.4
static3==0.5.1
wsgiref==0.1.2
